property_agent/pyproject.toml
[project]
name = "property-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.27.4",
    "opentelemetry-instrumentation-google-genai>=0.7b0",
]
